Change Decimal QL handling to hexadecimal
The displaying of the values in the app can be a bit confusing and
inconsistent - for example to emit external QL of 20, user has to input
value of 32, and examples given in README propose using QL of 11, at
which users complain. I propose following the spec - displaying and
configuring hex values.

Signed-off-by: Przemyslaw Korba <przemyslaw.korba@intel.com>