synce4l: fix sync on missing extended tlv
If the source port does not receive the extended tlv but SyncE device is
configured for such support, the incoming QL shall be validated with
the not-extended list of valid QLs, thus the priority and validity of
the source is correctly assigned.

Previously for described case the source was not considered as valid,
which is not compliant with the SyncE specification.

fix #38

Signed-off-by: Arkadiusz Kubalewski <arkadiusz.kubalewski@intel.com>