synce4l: fix duplicated pin data
Depending of the initialization order (driver registers dpll device and
pins vs synce4l load), it is possible that the port pin data is
initialized twice (once from config file, second from dpll pin dump).
This is due to association of a pin and netdevice which is not received
at the same time as the new pin notifications, also this comes from RT
netlink subsystem with different netlink family.

Remove the duplicated pin's data if it was already initialized and
assigned a data on device creation.

Signed-off-by: Arkadiusz Kubalewski <arkadiusz.kubalewski@intel.com>