synce4l: add internal device priority configuration
When building the source priority list for a device, and comparing each
two clock sources it is possible that multiple sources have the same
priority - the same QL (and extended QL) values.

Allow the user to explicitly assign an internal priority value for each
source within config file. Use the internal priority value for final
verdict when comparing two sources for their quality, previously the
compare function returned "left" source as better quality even if both
sources had the same priority, now the user is able to configure which
source is preferred in such case.

fix #17

Signed-off-by: Arkadiusz Kubalewski <arkadiusz.kubalewski@intel.com>