Fix DPLL pin ID assignment with invalid input
If an input was specified with invalid board label, the input would get
the pin ID of the next valid input and appear to be working. This was
caused by the dpll_mon_pin_id_update() function expecting that all
requested IDs are provided.

Add a flag to mark the pin which has the ID requested and update only
that pin to prevent the incorrect assignment.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>