Fix crash in dpll_rt_recv()
A crash was observed with E810 firmware 4.50 and kernel 6.11-rc1 in
dpll_rt_recv() trying to assign pin id to the pin which was removed by
remove_no_ifname_pin(). The pin is not a duplicate.

Add a third parameter to remove_no_ifname_pin() to avoid removing the
incorrect pin.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>